requirements-test.txt
fastapi==0.115.4
httpx==0.27.2
joblib==1.4.2
numpy==1.26.4
opencensus-ext-azure==1.1.13
pandas==2.2.3
pytest==8.3.3
uvicorn==0.32.0
scikit-learn==1.5.2
nltk==3.9.1